requirements.txt
Flask==2.0.1
exif==1.0.0
Pillow==8.2.0
